# apps/academic-service/src/modules/activity/activity.controller.ts
import { SetPermissions } from '@app/common/auth/decorators';
import { CacheTTL } from '@nestjs/cache-manager';
import {
  Body,
  Controller,
  Delete,
  Get,
  Param,
  Patch,
  Post,
  Query,
} from '@nestjs/common';
import { ActivityService } from './activity.service';
import { CreateActivityDto } from './dto/create-activity.dto';
import { FindAllQueryDto } from './dto/find-all-query.dto';
import { UpdateActivityDto } from './dto/update-activity.dto';

@Controller('activity')
@CacheTTL(1000 * 60 * 15) // 15 minutes
export class ActivityController {
  constructor(private readonly activityService: ActivityService) {}

  @SetPermissions('activities:add')
  @Post()
  create(@Body() createActivityDto: CreateActivityDto) {
    return this.activityService.create(createActivityDto);
  }

  @SetPermissions('activities:read')
  @Get()
  findAll(@Query() query: FindAllQueryDto) {
    return this.activityService.findAll({
      limit: query.limit || 10,
      offset: query.offset || 0,
    });
  }

  @SetPermissions('activities:read')
  @Get(':id')
  findOne(@Param('id') id: string) {
    return this.activityService.findOne(id);
  }

  @SetPermissions('activities:update')
  @Patch(':id')
  update(
    @Param('id') id: string,
    @Body() updateActivityDto: UpdateActivityDto,
  ) {
    return this.activityService.update(id, updateActivityDto);
  }

  @SetPermissions('activities:delete')
  @Delete(':id')
  remove(@Param('id') id: string) {
    return this.activityService.remove(id);
  }
}