# apps/docs-service/src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Academic Management Microservices API
  description: >-
    This is the unified API documentation for the Academic Management System, built using a production-grade microservices architecture. 
    All services are orchestrated via the Kong API Gateway, providing a single entry point for all client requests.

    The system is composed of several specialized services:
    - **User Service**: Handles identity management for students, teachers, and staff.
    - **Academic Service**: Manages course catalogs, semesters, scheduling, and notifications.
    - **Registry Service**: Governs financial transactions, fee management, and academic reporting.
    - **Messenger Service**: Facilitates real-time communication between system participants.

    Auth0 is utilized for secure, centralized authentication and authorization across the entire ecosystem.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api/v1
    description: Kong API Gateway (Local)
  - url: https://api.example.com/api/v1
    description: Production Gateway
...
security:
  - bearerAuth: []
paths:
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user data
      security:
        - bearerAuth: []
      responses:
        '200':
          description: user objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/profile:
    get:
      tags: [Authentication]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: user profile objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/roles:
    get:
      tags: [Authentication]
      summary: List users roles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of user roles objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /academic/semester/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Semesters]
      summary: Get single semester
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Semester'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Semesters]
      summary: Update semester
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateDTO'
            example:
              status: graduated
      responses:
        '202':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Semesters]
      summary: Delete semester
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /academic/course/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Courses]
      summary: Get course
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Courses'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Courses]
      summary: Update course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateDTO'
            example:
              semesterId: a1593306-7fd2-453b-a34c-79c165a9380b
              code: CS101
              name: Introduction to Computer Science
              description: Basic concepts of computing
              credits: 3
      responses:
        '202':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Courses'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Courses]
      summary: Delete course
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /academic/course/{courseId}/schedule/{id}:
    parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [CourseSchedules]
      summary: Get single schedules
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of schedules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseSchedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [CourseSchedules]
      summary: Create course schedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseScheduleUpdateDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseSchedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [CourseSchedules]
      summary: Create course schedule
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /academic/room/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Room]
      summary: Get room
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Room]
      summary: Update room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateDto'
      responses:
        '202':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Courses'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Room]
      summary: Delete room
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /academic/activity/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Activities]
      summary: Get single activity
      security:
        - bearerAuth: []
      responses:
        '200':
          description: activity record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Activities]
      summary: Update activity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityUpdateDTO'
      responses:
        '202':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Activities]
      summary: Delete activity
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MeResponse:
      type: object
      properties:
        app_name/roles:
          type: array
          items:
            type: string
          description: Custom claim containing assigned roles.
        iss:
          type: string
          description: Issuer identifier for the JWT.
          example: https://domain.com/
        sub:
          type: string
          description: Unique subject (user) identifier.
          example: auth0|6915a0beed*******
        aud:
          type: array
          items:
            type: string
          description: Audience values for which this token is intended.
        iat:
          type: integer
          format: int64
          description: Issued-at timestamp (Unix).
        exp:
          type: integer
          format: int64
          description: Expiration timestamp (Unix).
        scope:
          type: string
          description: OAuth scopes included in the token.
        azp:
          type: string
          description: Authorized party (client ID).
        permissions:
          type: array
          description: Permissions assigned to the user.
          items:
            type: string
...
    RoleData:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        name:
          type: string
      required: [id, role]
...
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        enrolmentDate:
          type: string
          format: date
        status:
          type: string
          default: 'ACTIVE'
        session:
          type: string
          default: 2021-22
        graduationYear:
          type: integer
          default: 2026
        currentSemesterId:
          type: string
          format: uuid

      required:
        - id
        - userId
        - status
        - session
        - currentSemesterId
...
    Semester:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]
...
    Courses:
      type: object
      properties:
        id:
          type: string
          format: uuid
        semesterId:
          type: string
          format: uuid
        code:
          type: string
          default: 2999
        name:
          type: string
          default: Introduction to Computer Science
        description:
          type: string
          description: Basic concepts of computing
        credits:
          type: integer
          format: int32
          default: 3
      required: [id, code, name]
    CourseCreateDTO:
      type: object
      required: [code, name, credits, semesterId]
      properties:
        semesterId:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        credits:
          type: integer
    CourseUpdateDTO:
      type: object
      properties:
        semesterId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        credits:
          type: integer

    # course generate
...
    CourseSchedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        dayOfWeek:
          type: string
          default: MONDAY
        startTime:
          type: string
          format: date
          description: time in HH:MM:SS
        endTime:
          type: string
          format: date
          description: time in HH:MM:SS
        room:
          type: string
          default: COM-201
        teacherId:
          type: string
          format: uuid
      required:
        - id
        - courseId
        - dayOfWeek
        - startTime
        - endTime
        - room
...
    CourseScheduleUpdateDTO:
      type: object
      properties:
        courseId:
          type: string
          format: uuid
        dayOfWeek:
          type: string
          default: MONDAY
        startTime:
          type: string
          format: date
          description: time in HH:MM:SS
        endTime:
          type: string
          format: date
          description: time in HH:MM:SS
        room:
          type: string
          default: COM-201
        teacherId:
          type: string
          format: uuid
    # room
    RoomType:
      type: string
      enum: [office, classroom, lab, washroom]
      description: The category/type of the room.
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          example: 'Advanced Robotics Lab'
        location:
          type: string
          example: 'Building B, 3rd Floor'
        type:
          $ref: '#/components/schemas/RoomType'
        capacity:
          type: integer
          minimum: 1
          example: 30
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - type
        - capacity
...
    RoomUpdateDto:
      type: object
      description: Partial update object for Room. All fields are optional.
      properties:
        name:
          type: string
        location:
          type: string
        type:
          $ref: '#/components/schemas/RoomType'
        capacity:
          type: integer
          minimum: 1
    # Attendance
...
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        activityType:
          type: string
          default: Notice board
        description:
          type: string
          default: description
        relatedEntityType:
          type: string
          default: entity type
        relatedEntityId:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
      required: [id, activityType]
...
    ActivityUpdateDTO:
      type: object
      required: [activityType]
      properties:
        activityType:
          type: string
          default: meeting
        description:
          type: string
          default: description
        relatedEntityType:
          type: string
          default: entity type
        relatedEntityId:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid

    # Notices
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          example:
            code: 400
            message: Bad Request
            fields:
              - field: field1
                values:
                  - field1 cannot be empty.
                  - field1 cannot be empty.

              - field: field2
                values:
                  - field2 must be at least 8 characters long.
                  - field2 must be at least 8 characters long.
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          example:
            code: 401
            statusCode: Unauthorized
    Forbidden:
      description: Forbidden
      content:
        application/json:
          example:
            statusCode: 403
            message: Forbidden resource
            error: Forbidden
    NotFound:
      description: Not Found
      content:
        application/json:
          example:
            statusCode: 404
            message: 'Not found!'
            error: 'Not found!'
...